third_party/blink/web_tests/fast/workers/worker-onerror-01-expected.txt

This tests that invalid syntax triggers 'worker.onerror'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Page-level worker.onerror handler triggered:
PASS errorEvent.message is "Uncaught SyntaxError: Unexpected token '}'"
PASS stripURL(errorEvent.filename) is "[blob: URL]"
PASS errorEvent.lineno is 2
PASS errorEvent.colno is 9
PASS errorEvent.error is null
PASS successfullyParsed is true

TEST COMPLETE
